+++ waterfall-cad/src/Waterfall/TwoD/Path2D.hs
@@ -26,6 +26,7 @@ module Waterfall.TwoD.Path2D
 , closeLoop2D
 , reversePath2D
 , splice2D
+, splitPath2D
 ) where 
 
 import Waterfall.TwoD.Internal.Path2D (Path2D(..))
@@ -154,3 +155,7 @@ reversePath2D = reversePath
 -- | `splice` with the type fixed to `Path2D`
 splice2D :: Path2D -> V2 Double -> (V2 Double, Path2D)
 splice2D = splice
+
+-- | `splitPath` with the type fixed to `Path2D`
+splitPath2D :: Path2D -> [Path2D]
+splitPath2D = splitPath